# swagger/swagger.yaml
consumes:
- application/json
definitions:
...
  Center:
    description: a center detail (default view)
    example:
      address:
        city: Adipisci et praesentium labore.
        country: Reiciendis eum natus rerum.
        created_at: "1970-06-21T17:34:49Z"
        id: Accusantium vel.
        lat: 0.4600049977027722
        lng: 0.35516983948280595
        name: Illum voluptatibus tenetur numquam eum.
        state: Est aut ducimus ducimus repellat minus.
        street1: Rerum non ipsa et.
        street2: Dolore labore sint qui.
        zip_code: Dolor in inventore perferendis molestiae vitae.
      address_id: Officia a.
      city:
        created_at: "1978-01-18T22:02:55Z"
        id: Officiis rerum qui amet expedita.
        name: Voluptatum quos aut accusantium.
      city_id: Vel ullam laudantium aliquam magnam atque reiciendis.
      created_at: "1997-07-08T20:24:22Z"
      id: Rerum et voluptatum rerum error eum quisquam.
      name: Quae voluptatibus reprehenderit provident.
    properties:
      address:
        $ref: '#/definitions/Address'
      address_id:
        example: Officia a.
        type: string
      city:
        $ref: '#/definitions/City'
      city_id:
        example: Vel ullam laudantium aliquam magnam atque reiciendis.
        type: string
      created_at:
        example: "1997-07-08T20:24:22Z"
        format: date-time
        type: string
      id:
        example: Rerum et voluptatum rerum error eum quisquam.
        type: string
      name:
        example: Quae voluptatibus reprehenderit provident.
        type: string
    title: 'Mediatype identifier: application/vnd.center+json; view=default'
    type: object
...
  Centers:
    description: A list of Center (default view)
    example:
      data:
      - address:
          city: Adipisci et praesentium labore.
          country: Reiciendis eum natus rerum.
          created_at: "1970-06-21T17:34:49Z"
          id: Accusantium vel.
          lat: 0.4600049977027722
          lng: 0.35516983948280595
          name: Illum voluptatibus tenetur numquam eum.
          state: Est aut ducimus ducimus repellat minus.
          street1: Rerum non ipsa et.
          street2: Dolore labore sint qui.
          zip_code: Dolor in inventore perferendis molestiae vitae.
        address_id: Officia a.
        city:
          created_at: "1978-01-18T22:02:55Z"
          id: Officiis rerum qui amet expedita.
          name: Voluptatum quos aut accusantium.
        city_id: Vel ullam laudantium aliquam magnam atque reiciendis.
        created_at: "1997-07-08T20:24:22Z"
        id: Rerum et voluptatum rerum error eum quisquam.
        name: Quae voluptatibus reprehenderit provident.
      paginate:
        current_page: 1.7282615592539546e+17
        total_item: 6.998965408074191e+17
        total_page: 5.952561025580748e+18
    properties:
      data:
        example:
        - address:
            city: Adipisci et praesentium labore.
            country: Reiciendis eum natus rerum.
            created_at: "1970-06-21T17:34:49Z"
            id: Accusantium vel.
            lat: 0.4600049977027722
            lng: 0.35516983948280595
            name: Illum voluptatibus tenetur numquam eum.
            state: Est aut ducimus ducimus repellat minus.
            street1: Rerum non ipsa et.
            street2: Dolore labore sint qui.
            zip_code: Dolor in inventore perferendis molestiae vitae.
          address_id: Officia a.
          city:
            created_at: "1978-01-18T22:02:55Z"
            id: Officiis rerum qui amet expedita.
            name: Voluptatum quos aut accusantium.
          city_id: Vel ullam laudantium aliquam magnam atque reiciendis.
          created_at: "1997-07-08T20:24:22Z"
          id: Rerum et voluptatum rerum error eum quisquam.
          name: Quae voluptatibus reprehenderit provident.
        items:
          $ref: '#/definitions/CenterGeneral'
        type: array
      paginate:
        $ref: '#/definitions/Paginate'
    title: 'Mediatype identifier: application/vnd.centers+json; view=default'
    type: object
  Cities:
    description: A list of City (default view)
    example:
      data:
      - created_at: "1997-12-21T02:00:44Z"
        id: Dolores ut.
        name: Illum amet suscipit excepturi excepturi.
      paginate:
        current_page: 1.7282615592539546e+17
        total_item: 6.998965408074191e+17
        total_page: 5.952561025580748e+18
    properties:
      data:
        example:
        - created_at: "1997-12-21T02:00:44Z"
          id: Dolores ut.
          name: Illum amet suscipit excepturi excepturi.
        items:
          $ref: '#/definitions/CityGeneral'
        type: array
      paginate:
        $ref: '#/definitions/Paginate'
    title: 'Mediatype identifier: application/vnd.cities+json; view=default'
    type: object
  City:
    description: a city detail (default view)
    example:
      created_at: "1978-01-18T22:02:55Z"
      id: Officiis rerum qui amet expedita.
      name: Voluptatum quos aut accusantium.
    properties:
      created_at:
        example: "1978-01-18T22:02:55Z"
        format: date-time
        type: string
      id:
        example: Officiis rerum qui amet expedita.
        type: string
      name:
        example: Voluptatum quos aut accusantium.
        type: string
    title: 'Mediatype identifier: application/vnd.city+json; view=default'
    type: object
...
  Login_response:
    description: Login response (default view)
    example:
      token: Impedit vero iste voluptas et exercitationem.
      user_info:
        company_address: Nobis ab quia pariatur corrupti laborum est.
        email: Eos saepe minima laborum tempora dolores.
        name: Dolorem esse provident nisi sit eius est.
        phone: Quis consequatur.
    properties:
      token:
        example: Impedit vero iste voluptas et exercitationem.
        type: string
      user_info:
        $ref: '#/definitions/User_info'
    title: 'Mediatype identifier: application/vnd.login_response+json; view=default'
    type: object
  LoginPayload:
    example:
      email: jamesbond@gmail.com
      password: abcd1234
    properties:
      email:
        example: jamesbond@gmail.com
        format: email
        maxLength: 400
        minLength: 6
        type: string
      password:
        example: abcd1234
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - password
    title: LoginPayload
    type: object
...
  RegisterPayload:
    example:
      company_address: 445 Mount Eden Road, Mount Eden, Auckland
      company_name: Express company
      contact_name: Doe
      email: jamesbond@gmail.com
      password: abcd1234
      phone: e46tmo2z7y
    properties:
      company_address:
        example: 445 Mount Eden Road, Mount Eden, Auckland
        maxLength: 500
        minLength: 1
        type: string
      company_name:
        example: Express company
        maxLength: 500
        minLength: 1
        type: string
      contact_name:
        example: Doe
        maxLength: 200
        minLength: 1
        type: string
      email:
        example: jamesbond@gmail.com
        format: email
        maxLength: 150
        minLength: 6
        type: string
      password:
        example: abcd1234
        maxLength: 100
        minLength: 8
        type: string
      phone:
        example: e46tmo2z7y
        maxLength: 50
        type: string
    required:
    - email
    - password
    - contact_name
    - phone
    - company_address
    - company_name
    title: RegisterPayload
    type: object
...
  Schedules:
    description: A list of Schedule by Center ID (default view)
    example:
      data:
      - center:
          address:
            city: Adipisci et praesentium labore.
            country: Reiciendis eum natus rerum.
            created_at: "1970-06-21T17:34:49Z"
            id: Accusantium vel.
            lat: 0.4600049977027722
            lng: 0.35516983948280595
            name: Illum voluptatibus tenetur numquam eum.
            state: Est aut ducimus ducimus repellat minus.
            street1: Rerum non ipsa et.
            street2: Dolore labore sint qui.
            zip_code: Dolor in inventore perferendis molestiae vitae.
          address_id: Officia a.
          city:
            created_at: "1978-01-18T22:02:55Z"
            id: Officiis rerum qui amet expedita.
            name: Voluptatum quos aut accusantium.
          city_id: Vel ullam laudantium aliquam magnam atque reiciendis.
          created_at: "1997-07-08T20:24:22Z"
          id: Rerum et voluptatum rerum error eum quisquam.
          name: Quae voluptatibus reprehenderit provident.
        center_id: Ut qui qui dolorem.
        end_at: "1975-07-07T07:29:52Z"
        id: Possimus sed porro.
        start_at: "1995-12-27T22:48:41Z"
      paginate:
        current_page: 1.7282615592539546e+17
        total_item: 6.998965408074191e+17
        total_page: 5.952561025580748e+18
    properties:
      data:
        example:
        - center:
            address:
              city: Adipisci et praesentium labore.
              country: Reiciendis eum natus rerum.
              created_at: "1970-06-21T17:34:49Z"
              id: Accusantium vel.
              lat: 0.4600049977027722
              lng: 0.35516983948280595
              name: Illum voluptatibus tenetur numquam eum.
              state: Est aut ducimus ducimus repellat minus.
              street1: Rerum non ipsa et.
              street2: Dolore labore sint qui.
              zip_code: Dolor in inventore perferendis molestiae vitae.
            address_id: Officia a.
            city:
              created_at: "1978-01-18T22:02:55Z"
              id: Officiis rerum qui amet expedita.
              name: Voluptatum quos aut accusantium.
            city_id: Vel ullam laudantium aliquam magnam atque reiciendis.
            created_at: "1997-07-08T20:24:22Z"
            id: Rerum et voluptatum rerum error eum quisquam.
            name: Quae voluptatibus reprehenderit provident.
          center_id: Ut qui qui dolorem.
          end_at: "1975-07-07T07:29:52Z"
          id: Possimus sed porro.
          start_at: "1995-12-27T22:48:41Z"
        items:
          $ref: '#/definitions/ScheduleGeneral'
        type: array
      paginate:
        $ref: '#/definitions/Paginate'
    title: 'Mediatype identifier: application/vnd.schedules+json; view=default'
    type: object
  Token:
    description: A token (default view)
    example:
      token: Beatae ipsum.
    properties:
      token:
        description: A JWT token
        example: Beatae ipsum.
        type: string
    title: 'Mediatype identifier: application/vnd.token+json; view=default'
    type: object
...
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609066e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609066e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
host: localhost:9000
info:
  contact:
    email: hieu.phq@gmail.com
    name: Hieu Phan
  description: An API for califit-be
  title: The califit-be Backend API
  version: ""
paths:
  /api/auth/login:
    post:
      description: Sign a company user in
      operationId: authentication#login
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/LoginPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.login_response+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Login_response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: login authentication
      tags:
      - authentication
  /api/auth/logout:
    post:
      description: Sign a company user in
      operationId: authentication#logout
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/LoginPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.login_response+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Login_response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: logout authentication
      tags:
      - authentication
  /api/auth/register:
    post:
      description: Create a new user
      operationId: authentication#register
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/RegisterPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.token+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
      schemes:
      - https
      summary: register authentication
      tags:
      - authentication
  /api/center:
    get:
      description: List centers
      operationId: center#list
      parameters:
      - default: ""
        description: query for find centers in city
        in: query
        name: cityID
        required: false
        type: string
      - default: 10
        description: limit for paginate
        in: query
        name: limit
        required: false
        type: integer
      - default: ""
        description: query for find centers by name
        in: query
        name: name
        required: false
        type: string
      - default: 0
        description: offset for paginate
        in: query
        name: offset
        required: false
        type: integer
      produces:
      - application/vnd.centers+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Centers'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: list center
      tags:
      - center
  /api/center/{centerID}:
    get:
      description: Get a center by ID
      operationId: center#show
      parameters:
      - description: Center ID
        in: path
        name: centerID
        required: true
        type: string
      produces:
      - application/vnd.center+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Center'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: show center
      tags:
      - center
  /api/city:
    get:
      description: List cities
      operationId: city#list
      parameters:
      - default: 10
        description: limit for paginate
        in: query
        name: limit
        required: false
        type: integer
      - default: ""
        description: query for find cities by name
        in: query
        name: name
        required: false
        type: string
      - default: 0
        description: offset for paginate
        in: query
        name: offset
        required: false
        type: integer
      produces:
      - application/vnd.cities+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Cities'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: list city
      tags:
      - city
  /api/city/{cityID}:
    get:
      description: Get a city by ID
      operationId: city#show
      parameters:
      - description: City ID
        in: path
        name: cityID
        required: true
        type: string
      produces:
      - application/vnd.city+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/City'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: show city
      tags:
      - city
  /api/schedule/{centerID}:
    get:
      description: List schedule
      operationId: schedule#list
      parameters:
      - description: Center ID
        in: path
        name: centerID
        required: true
        type: string
      - default: 10
        description: limit for paginate
        in: query
        name: limit
        required: false
        type: integer
      - default: 0
        description: offset for paginate
        in: query
        name: offset
        required: false
        type: integer
      produces:
      - application/vnd.goa.error
      - application/vnd.schedules+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Schedules'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: list schedule
      tags:
      - schedule
...
produces:
- application/json
...
schemes:
- https
securityDefinitions:
  jwt:
    description: |-
      Use JWT to authenticate

      **Security Scopes**:
        * `api:access`: API access
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"